+++ RelationalQueryEvaluator/PhysicalOperatorVisitor.cpp
@@ -620,7 +620,7 @@ endOfOuterCycle:
 
 void SortResolvingPhysicalOperatorVisitor::visitMergeEquiJoin(MergeEquiJoin * node)
 {
-	map<ulong, ulong> pairs;
+	map<int, int> pairs;
 	vector<shared_ptr<Expression> > condition;
 	AlgebraCompiler::serializeExpression(node->condition, condition);
 	for (auto it = condition.begin(); it != condition.end(); ++it)
@@ -753,7 +753,7 @@ void SortResolvingPhysicalOperatorVisitor::visitHashAntiJoin(HashAntiJoin * node
 
 void SortResolvingPhysicalOperatorVisitor::visitMergeAntiJoin(MergeAntiJoin * node)
 {
-	map<ulong, ulong> pairs;
+	map<int, int> pairs;
 	vector<shared_ptr<Expression> > condition;
 	AlgebraCompiler::serializeExpression(node->condition, condition);
 	for (auto it = condition.begin(); it != condition.end(); ++it)
